bug reports 2.0

+++ C#/Calculator/Number.cs
@@ -6,10 +6,10 @@ namespace Calculator
     {
         internal Number(string num,string character)
         {
-            Num = float.Parse(num,new CultureInfo("en-US"));
+            Num = decimal.Parse(num,new CultureInfo("en-US"));
             Character = character;
         }
-        internal float Num { get; set; }
+        internal decimal Num { get; set; }
         internal string Character { get; set; }
     }
 }
